requirements.txt
asgiref==3.9.1
click==8.3.0
colorama==0.4.6
dj-config-url==0.1.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
insultgenerator==1.0.3
loremipsum==1.0.5
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.36.0
White-Noise==0.1.0
whitenoise==6.10.0
